debugger: rename internal module
All other files in `lib/internal` do not start with `_`. For
consistency, move `lib/internal/debugger/_inspect.js` to
`lib/internal/debugger/inspect.js`.

PR-URL: https://github.com/nodejs/node/pull/39378
